backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.6.1
sqlalchemy==2.0.36
docker==7.1.0
psutil==6.1.0
pyyaml==6.0.2
httpx==0.28.0
python-multipart==0.0.17
